+++ src/position.cc
@@ -16,7 +16,7 @@
 namespace ccls {
 Pos Pos::fromString(const std::string &encoded) {
   char *p = const_cast<char *>(encoded.c_str());
-  uint16_t line = uint16_t(strtoul(p, &p, 10) - 1);
+  int32_t line = int32_t(strtoul(p, &p, 10) - 1);
   assert(*p == ':');
   p++;
   int16_t column = int16_t(strtol(p, &p, 10)) - 1;
@@ -32,14 +32,14 @@ std::string Pos::toString() {
 Range Range::fromString(const std::string &encoded) {
   Pos start, end;
   char *p = const_cast<char *>(encoded.c_str());
-  start.line = uint16_t(strtoul(p, &p, 10) - 1);
+  start.line = int32_t(strtoul(p, &p, 10) - 1);
   assert(*p == ':');
   p++;
   start.column = int16_t(strtol(p, &p, 10)) - 1;
   assert(*p == '-');
   p++;
 
-  end.line = uint16_t(strtoul(p, &p, 10) - 1);
+  end.line = int32_t(strtoul(p, &p, 10) - 1);
   assert(*p == ':');
   p++;
   end.column = int16_t(strtol(p, nullptr, 10)) - 1;
@@ -47,9 +47,9 @@ Range Range::fromString(const std::string &encoded) {
 }
 
 bool Range::contains(int line, int column) const {
-  if (line > UINT16_MAX)
+  if (line > INT32_MAX)
     return false;
-  Pos p{(uint16_t)line, (int16_t)std::min<int>(column, INT16_MAX)};
+  Pos p{(int32_t)line, (int16_t)std::min<int>(column, INT16_MAX)};
   return !(p < start) && p < end;
 }
